refactor(web-serial): pi zero session api and bootstrap split
Made-with: Cursor
